chore: useHotkeys, Root, NextAuth
